ProductServiceImplにて削除処理時のquantityのゼロ判断を実装

+++ src/main/java/com/raisetech/inventoryapi/service/ProductServiceImpl.java
@@ -1,18 +1,23 @@
 package com.raisetech.inventoryapi.service;
 
+import com.raisetech.inventoryapi.entity.InventoryProduct;
 import com.raisetech.inventoryapi.entity.Product;
 import com.raisetech.inventoryapi.exception.ResourceNotFoundException;
+import com.raisetech.inventoryapi.mapper.InventoryProductMapper;
 import com.raisetech.inventoryapi.mapper.ProductMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductServiceImpl implements ProductService {
     private final ProductMapper productMapper;
+    private final InventoryProductMapper inventoryProductMapper;
 
-    public ProductServiceImpl(ProductMapper productMapper) {
+    public ProductServiceImpl(ProductMapper productMapper, InventoryProductMapper inventoryProductMapper) {
         this.productMapper = productMapper;
+        this.inventoryProductMapper = inventoryProductMapper;
     }
 
     @Override
@@ -39,6 +44,17 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void deleteProductById(int id) {
         productMapper.findById(id).orElseThrow(() -> new ResourceNotFoundException("resource not found with id: " + id));
-        productMapper.deleteProductById(id);
+        Optional<InventoryProduct> inventoryProductOptional = inventoryProductMapper.findInventoryByProductId(id);
+        if (inventoryProductOptional.isPresent()) {
+            InventoryProduct inventoryProduct = inventoryProductOptional.get();
+            int quantity = inventoryProduct.getQuantity();
+            if (quantity == 0) {
+                productMapper.deleteProductById(id);
+            } else {
+                throw new IllegalStateException("Cannot delete Product because the quantity is 0");
+            }
+        } else {
+            productMapper.deleteProductById(id);
+        }
     }
 }
